fastapi_app/requirements.txt
httpx==0.26.0
fastapi==0.108.0
uvicorn==0.25.0
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp==1.24.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-langchain==0.17.2
langchain
langchain-openai
python-dotenv
loguru
faiss-cpu # memory vectorstore
beautifulsoup4 # wev retrieval